+++ src/main/java/com/mike/lunchvoter/util/Constants.java
@@ -5,6 +5,6 @@ import lombok.experimental.UtilityClass;
 @UtilityClass
 public class Constants {
 
-    public static final int NON_EXISTING_ID = -1;
+    public static final Long NON_EXISTING_ID = -1L;
 
 }
